// models/MemberModel.tsx
export interface MemberModel {
    id: number;
    firstName: string;
    lastName: string;
    age: number;
    gender: string;
    email: string;
    phone: string;
    birthDate: string;
    image: string;
    bloodGroup: string;
    height: number;
    weight: number;
    macAddress: string;
    university: string;
}